bzl: bump rules_go to 0.40.1 (#54912)
Bump rules_go to a newer version. I didn't go for `0.41.0` because it
introduces a breaking change with how we deal with the googleapis and
that will require some work to patch the back compat tests. So I'd
rather bump this, update to go 1.20.5 and then deal with the 0.41.0
update :)

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

CI