embeddings: get rev from index instead of from job
I noticed this while thinking about inconsistencies between the job
table and indexes in the store.

This hasn't led to problems yet, but it very well might if the jobs
table doesn't reflect the state of the store anymore.

We should not consider the job table as authorative when it comes to
index meta. Instead we should rely on the store or at least a db table that
is eventually consistent with the store.