embeddings: name embeddings by id not by name (#54782)
With this PR we migrate embeddings to the new naming schema `<repo
ID>.embeddingindex`. Old embeddings can still be read. New embeddings
(full index and incremental) will follow the new naming schema.


Example: repo "qa-repo".

Old
```
qa_repo_d5e73055b0ac12179e03a9e6dc330d1c.embeddingindex
```

New
```
14.embeddingindex
```
*Why?*

- Our current naming schema makes it very inefficient to map embeddings
to their repo. However, we need such a mapping, because we want to
establish a representation of the existing embeddings in our database.
So far we used the `repo_embeddings_jobs` table as a proxy, which is
prone to diverge from the state of the store at some point. Using the
new naming schema makes it trivial to map embeddings to their respective
repos.
- Once we have embeddings represented in the db , we can leverage unions
with the repo table to unlock improvements for admins, such as filtering
the repos in **site-admin > repositories** based on whether they have
embeddings or not.

## Test plan
- updated unit test
- manual testing: I verified that 
  - existing embeddings can stiill be read
  - existing embeddings are used for incremental updates
- embeddings based on name are deleted once embeddings based on ID have
been created