licensing: allow multiple instances with same license key
Rationale: https://docs.google.com/document/d/1o8fypurDxsukEbn9GaPQDv96PNyBPcoKlxpBQTW6l7U/edit

We do have customers who were broken because they use more than one instance
with the same license key. We changed the v5.1 license key validation to
only allow 1 license key per 1 instance. Until v5.1 it was a common setup to have
multiple instances with the same key and it adds burden to customers and CE/TA
teams to fix this.

This change is meant to switch from being reactive - mark the license on the instance
as invalid when such a situation happens to be proactive - send a slack message to us instead
and notify the customer that we need to change their license keys.