licensing: display license validity status in site-admin license page (#54961)
This PR updates the UI for the License page `/site-admin/license` so it displays more information about the validity of a license key.

#### Before

<img width="1117" alt="CleanShot 2023-07-14 at 13 10 39@2x" src="https://github.com/sourcegraph/sourcegraph/assets/25608335/55a458a2-b26b-4d7b-9b0c-07c34cbaf785">

#### After

<img width="1124" alt="CleanShot 2023-07-14 at 13 09 54@2x" src="https://github.com/sourcegraph/sourcegraph/assets/25608335/44563423-6ddc-41f5-8d18-2500a33bd223">

## Test plan

<!-- All pull requests REQUIRE a test plan: https://docs.sourcegraph.com/dev/background-information/testing_principles -->
Manually updating the values of:

* `licensing:is_license_valid`
* `licensing:license_invalid_reason`

The first key should reflect the validity status of the Sourcegraph license key, while the second should return the reason it's invalid (if it is).

@sourcegraph/design I appreciate some feedback on the design here.

Co-authored-by: Erik Seliger <erikseliger@me.com>
(cherry picked from commit d29c89ccddee0c130fa30c25357e8a293012444a)