Embeddings: optimize memory usage during decode (#54972)
This makes memory usage ~constant during the decode process. Previously,
we were undersizing the large allocation (which caused it to reallocate
~10 extra times) for the index and making a bunch of additional smaller
allocations per loop. This meant we were spiking memory usage during
decoding, and even if the garbage collector could keep up (which I
wouldn't expect it to), we'd still need 2x the index size at the peak
for live memory. This was causing an OOM.

This PR accounts for a 2.2x reduction in allocated bytes and a 1.4x reduction in
allocation count. This is quite significant when our indexes are many
GBs in memory.