search-blitz: add attribution search
We are making improvements to attribution search performance and it
would be useful to start measuring it over time.

This commit extends the graphql client to do attribution search. We then
extend the QueryConfig struct such that if Snippet is specified we do
attribution search instead of the normal search. Because attribution
search is on arbitrary snippets, we read in files from the attribution
directory.

This commit also did some minor refactoring of graphql related code. We
move the graphql queries to a file to improve legibility. Additionally
the shared graphql code is pulled into a function to be shared.

Test Plan: test in prod