internal/gitserver: Make repoAddressCache manage expiry with jitter
We want to spread the cache expiry to avoid a spike in DB reads for
workloads that process repositories in bulk. As a result to support the
jitter we we make two behavioral changes to the repoAddressCache in this
commit:

1. The item now stores an expiresAt instead of createdAt
2. The Read API now returns `nil` if the item has expired by comparing
the current time with the expiresAt attribute of the item

Coming back to the jitter, we now add a jitter of ± 30 seconds (3%) to the
cache expiry time of 15 minutes. This helps spread the cache expiry and
flatten the workload on the DB.