gitserver: Don't delete repos when DB is down
First, this is not a terribly huge problem, as this cleanup job is off by default. But while debugging something else, this struck me as potentially dangerous.
In case of a DB outage or broken query, we would drain the entire gitserver disk here slowly because repo will be nil when general connectivity issues occur.
To fix this, we add an error check and only do this for not found errors.
I also dropped a duplicative log message. The cleanup function returning an error already logs all the relevant details.