gitserver: change the way repo name is checked for git command recording.
Previously, we checked for a substring, but it breaks in cases where given repo name is a prefix for another repo names, e.g.: "repo/a" is included, but "repo/ab" and "repo/a-1" are included too, because of the check.

This commit changes this behaviour to check for the suffix instead, which is not perfect, but judging from local testing and how the repos are named ("orgName/repoName"), this check will filter out most repos which were previously included by mistake.

Precise implementation requires a lot of effort to propagate repo names through the recording stack and can be done later.

Test plan:
Local sg run and test that repos with matching prefix are not recorded.