Fuzzy finder: allow line number at the end of a filename (#55064)
Previously, it was not possible to navigate straight to a specific line
number with the "Files" tab of the fuzzy finder. This commit adds the
ability to specify a line number with a `:NUMBER` suffix. This feature
makes it possible to copy-paste error messages that use the
`FILE:LINE:COLUMN` format and quickly open the location with the fuzzy
finder.

## Test plan

- New storybook test
- New unit tests for query parsing
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->