database: re-enable flaky test & remove sleep
This fixes #46510 by re-enabling the test and fixing its flakiness.

Test was flaky because it was `t.Parallel` but used `conf.Mock`, which
writes to a global var.

So I removed `t.Parallel`, which fixes the data race.

But I also removed the `time.Sleep`, beacuse we don't need that. We can
just reset the time.