site-admin: fix spinner on repositories page
This fixes an anoying behavior of the spinner on the Site-Admin >
Repositories page.

Current:

The spinner instantly shows up while we are fetching content. This is
fine for very very large instances, where fetching takes time, but on
instances like s2 this just leads to a flickering repositories count in
the top-left corner.

Fix:
We move the spinner down to the top of the list and only show it if
loading takes longer than 300ms.