Smart search: handle fully qualified symbol names
Previously, the query `type:symbol com.example.Main` didn't return
results even if there was a symbol that has the fully qualified name
`com.example.Main`. In our symbol backend, we store the "containerName"
`com.example` and "name" `Main` separately so it should be technically
possible to find the symbol by fully qualified name.

This commit fixes the problem by adding a smart search rule that takes
fully qualified symbol names and infers the correct syntax to search for
the container name. The current implementation only works for some
languages where the container name is declared in the contents of the
file. Ideally, the search syntax should have a filter for containerName
directly so that we can target it precisely including in languages like
Go or Python where the package name is implicit from the file path.