local: hotfix ctags mess with symbols (#55196)
This stuff is really messy, but this will at least unblock folks using
ctags command.

@davejrt is going to make this bunch of scripts completely irrelevant by
getting Bazel to fetch the universal ctags binary as part of our Q3
effort to rework local env to use Bazel more.

So, with that in mind, this PR is really the dumbest fix possible for
the time being. Honnestly these scripts are messy, and should be deleted
entirely anyway.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Locally tested.