move files to avoid bad import rule
the worker uses database.DB, but it is forbidden to import
that from cmd/ directories. Because the worker lived in the
licensing package before, licensing package started being
dependent on db, which it previously was not.

This caused dev/check/go-dbconn-import.sh to fail

Moving the files to a different package within worker
space has helped without affecting the functionality.