fix: move licensing checks to worker process (#54854)
Moving licensing checks to a singleton worker job. This way we make sure
that only one license check API call is being made on each 12-hour tick.

Previously this lived in a frontend service. If multiple frontend
services were running, this would fire multiple requests, which is
undesirable from many points of view - waste of resources, waste of
bandwidth, potential concurrency issues.

Having it fire multiple times would also break an assumption that we
have in the license check handler on dotcom. For checking if same
site_id is used on multiple instances, we assume that the time
difference between license check calls would be roughly 10-12 hours.
Calling twice in the same second breaks this assumption.

## Test plan

Tested locally that the worker starts