dx: draft the overview of the Bazel-client configuration (#54743)
## Context

Draft of the Bazel-client configuration overview based on 1-1 sessions
with Vova and Tom. It aims to cover all configuration pieces essential
for maintaining the Bazel integration with Javascript tooling and our
client packages. It starts with covering JS-specific parts in WORKSPACE
and root BUILD files and goes into more specific/complex aspects of our
toolset.

Sharing it early for feedback. Going to fill out the second half and
link it properly to the existing table contents in a follow-up PR.

## Test plan
N/a — docs


Co-authored-by: Jean-Hadrien Chabran <jh@chabran.fr>
Co-authored-by: Vova Kulikov <vova@sourcegraph.com>