source: truncate hostname for repos regardless of how many slashes contained (#55154)
Closes [#54975](https://github.com/sourcegraph/sourcegraph/issues/54975)

I made an assumption here that the repository name will always be in the format `<hostname>/...`. That way we can assume that the repo name in the `displayRepoName` function will always contain at least one slash.

This P

| Before  | After  |
|---|---|
| <img width="954" alt="CleanShot 2023-07-20 at 13 58 30@2x" src="https://github.com/sourcegraph/sourcegraph/assets/25608335/79618455-cfee-48c2-9830-5dc6d68f104e">  | <img width="940" alt="CleanShot 2023-07-20 at 13 57 50@2x" src="https://github.com/sourcegraph/sourcegraph/assets/25608335/5e96750c-a3d5-4e26-a759-f89abe2ed277">  |

## Test plan

<!-- All pull requests REQUIRE a test plan: https://docs.sourcegraph.com/dev/background-information/testing_principles -->

Add a mix of repos to external services with multiple amounts of slashes. In my case, I added Gerrit repos in the `gerrit.sgdev.org` namespace.

```json
"GERRIT": [
    {
      "url": "https://gerrit.sgdev.org",
      "username": "bolaji.o",
      "password": "C/Y4afoDftfYr19aySlluPFbVXdDLLhOGPcdjGzBlw",
      "projects": [ // If not set, all projects on the Gerrit instance will be mirrored
        "src-cli",
        "gabe",
        "gabe/test",
        "beatrix",
        "a/gabe/repo"
      ]
    }
  ]
```

The repository names will be displayed properly without the instance name.