[Backport 5.1] Implement experimental Azure OpenAI provider for completions and embeddings (#55178)
This adds experimental support for Azure OpenAI as an embeddings and
completions provider.

I added some experimental docs for it, and ran through it with a bunch
of repos and queries, seems to be on-par with the OpenAI providers
mostly. Only thing is that we make more requests for embeddings, because
Azure doesn't yet support the batch API.

Verified manually using our Azure deployment.

(cherry picked from commit 9a01c364b1810b1a62027c77f0b45dd7f6efbc17)