file sidebar: don't hide error by turning last file into folder
I was looking into #41750 and had to laugh out loud when I got to the
last file in the sidebar, though "huh, why is that one a folder?",
opened it and was greeted with an error message.

This here fixes the issue by not treating the error message as a child
of the previous entry, which turns the previous entry into a folder.