[gitlab] Add alternate way to confirm GitLab project visibility (#54426)
Original Slack thread
[here](https://sourcegraph.slack.com/archives/C05EMJM2SLR/p1687969384430849)

## Before this PR

We use a `min_access_level=20` check on GitLab project queries, because
`access_level` 10 is GUEST, and users with GUEST access cannot see repo
contents, so we don't want to give access to those repos on Sourcegraph
HOWEVER, for Internal repos on GitLab, users can see the contents of the
repo depending on how the repo is configured, regardless of their access
level. So this puts us in a conundrum. If we have `min_access_level=20`,
users can't see repos they should be able to see, but if we remove it,
users can see repos they shouldn't be able to see

## After this PR

Adds a feature flag named "gitLabProjectVisibilityExperimental". When
set to true:

@indradhanush and I diffed some responses from the GitLab API, depending
on whether or not you are part of the GitLab group or not, whether the
project is visible or not, etc. etc. and there is a common field:
`default_branch` is only visible in the project API response if the user
is able to see the contents of the project. If they cannot see the
contents of the project, they cannot see the default branch of the
project.
So by removing `min_access_level=20` and checking for the presence of
`default_branch` on the returned project list, we can determine whether
or not the user should be able to see the repo contents or not

## Test plan

VCR tests should be updated

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

(cherry picked from commit 8cb61f8630e7a74e5bc69fa7b856b2767cd5107d)