gRPC: sanitize utf8 in searcher (#55369)
This sanitizes sources of the `Content` field of `ChunkMatch` to ensure
that it is always valid UTF-8.

The underlying issue is that searcher does not exclude utf8 files when
searching, so we can't guarantee that the contents are valid utf8.

(cherry picked from commit fabb0e9a39735f6e967c1937250331c260ae9edf)