Add sg commands to build wolfi packages and images locally (#55316)
As part of improving the Wolfi dev experience, we want to be able to
build packages and images locally using sg. It should also be easy for
developers to refer to packages they have built locally in their images,
for testing. Currently this is difficult due to repo signatures and
docker mount points.

https://github.com/sourcegraph/security/issues/563

## Test plan

- Manually test commands
- CI

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->