oob: Always run enterprise OOB migrations in worker again (#55368)
In
https://sourcegraph.sourcegraph.com/github.com/sourcegraph/sourcegraph/-/commit/591d702d290c5a998bc08c4e4e66c8914407d542,
we moved this init script from enterprisecmd to frontend, which disabled
the enterprise migrations to be considered in other services - namely
worker which is running the migrations in the background for active
instances.

That means that we haven't been running a few migrations while instances
are running in 5.1, but that shouldn't be a big problem luckily because
- none are deprecated in 5.1
- none were introduced in 5.1
- only a few from 5.0 would potentially need to run
- but they likely already completed while the instance was on 5.0,
unless there was a lot of stuff to be migrated
- when an instance first booted on 5.1, these would have never run, but
also no old data would exist

## Test plan

Verified migrations run again.

(cherry picked from commit e2333de2bdff310cddb6d74533fb5ecd11b2f051)