Add site admin banner warning when exceeding Cody limits
This PR adds a mechanism that syncs the rate limits in the background, stores the usage in redis every 30 minutes, and displays a banner warning when the rate limits are being exceeded:
- Info at 75%+
- Warning at 90%+
- Error at 100%

This PR also moves the code for talking to Cody Gateway to get usage data into the common codebase out of app, so that we get all the data exposed through graphql for every instance. This may be useful for debugging, or implementing additional APIs to visualize limits.