Backend: remove unnecessary internal API for external services (#55450)
This removes an unnecessary internal endpoint. Basically,
`schema.ClientConfiguration` is the only consumer of `conf.*Configs()`,
which are the only consumers of `internalClient.ExternalServiceConfigs`,
which just hits the database and pulls the URL off the config.

So, rather than calling an internal API from `frontend`, which is just
served by `frontend`, this PR updates `schema.ClientConfiguration` to
just hit the database directly for this information, allowing us to get
rid of one of the few remaining internal APIs.

This is an API that was slated to be converted to gRPC, so I'd rather
not convert things we can just delete.

(cherry picked from commit 7265e0428e0854bb5ff41004a77c1d06dfac5efc)