sveltekit: Add support for feature flags (and improve testing)
This PR adds a new store to make feature flags available throughout the
app. Unlike the implementation in the web app, this implementation
fetches all user flags when the page loads and updates them in
intervals.

In the future we might have to add more functionality to force update
the flags, but we can add those as needed.

This PR also updates the testing setup to make testing simpler and more
deterministic:

- Faker is now automatically initialized with a fixed reference date.
- Functionality for mocking feature flags has been added
- Added helpers for using fake timers with a fixed reference/system
  date (it turned out that enabling fake timers by default doesn't work
  well when testing user interaction).