sveltekit: Add badge wildcard component
This PR adds the bade wildcard component to the prototype.

Whenever styles should be "shareable" it's easier to use CSS modules
rather than Svelte `<style>` tags.
Similar to Button, this component allows to customize the DOM structure
via a "custom" slot which gets passed the generated class names.
(whether or not that's a good approach remains to be seen)

Additionally customization could als be done by using `badeClassName`
directly.