Allow embeddings job to exclude failed files from the index (#55180)
When a text input is submitted for generating embeddings the response
may be null. If we attempt retries and still cannot generate embeddings
for this input text then we return an error which calls for failing the
entire embed repo job.

[Slack
thread](https://sourcegraph.slack.com/archives/C053L1AQ0BC/p1688676751106069)

[Issue](https://github.com/sourcegraph/sourcegraph/issues/55469)

This PR introduces a configuration `ExcludeChunkOnError`. When set to
true an embed repo job will proceed with the rest of the embed repo job
when these generate embeddings errors occur. However, the file that
generated the input text which received an error is excluded from the
index as to avoid partially indexing the file.

I'll add more details on the first iteration of this solution and the
trade offs in a separate comment.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Embed test cases added

(cherry picked from commit b8e31fde275d155aed5fe4cea30f3514bd816402)